updated the css and reduced the number of fetch in js

+++ mkadi/test/game-stop/index.js
@@ -1,30 +1,40 @@
 let base_url =` http://localhost:3000/games`
-function myGames(games){
+
+document .addEventListener("DOMContentLoaded",() => {
+    fetchGames();
+    document.getElementById("gameForm").addEventListener('submit', addGame);
+})
+
+function fetchGames(){
+    fetch(base_url)
+    .then(res => res.json())
+    .then(data =>{
+        console.log("games are fetched:",data);
+        displaGames(data);
+   })
+   .catch(err => console.error("error occured fetching game:",err));
+}
+function displaGames(games){
     let gameLibrary = document.getElementById('gameinventory')
-    gameLibrary.innerHTML = ""
-    games.forEach(game =>{  let html = `
+    gameLibrary.innerHTML = "";
+
+    games.forEach(game =>{
+          let gameElement =`
         <div id="single-game" data-id="${game.id}">
             <img src="${game.image}" alt="${game.title}">
             <div>${game.description}</div>
             <div>${game.price}</div>
             <div>
-            <button class="delete" onclick="deleteGame(${game.id})">remove<button>
-            <button class="buy" onclick="postgame(${game.id})">download<button>
+            <button class="delete" onclick="deleteGame(${game.id})">Remove</button>
+            <button class="buy" onclick="postGame(${game.id})">Download</button>
             </div>
         </div>   
         `
-        gameLibrary.innerHTML += html
+        gameLibrary.innerHTML +=gameElement
     })
 }
 
-fetch(base_url)
-.then(res =>res.json())
-.then(data => {
-     console.log(data);
-     myGames(data);})
-.catch(err => console.log( "error occured in fetching game",err))
-
-document.getElementById('gameForm').addEventListener('submit',function(event){
+function addGame(event){
     event.preventDefault();
     let gameTitle = document.getElementById('title').value;
     let gameImage = document.getElementById('image').value;
@@ -46,25 +56,30 @@ fetch(base_url,{
 })
     .then(res => res.json())
     .then(data => {console.log("game added",data);
-     fetch(base_url)
-    .then(res => res.json())
-    .then(updatedData => displayGames(updatedData));
-})
-    .catch(err => console.log("error occured while adding game",err));
+     fetchGames();
+     document.getElementById("gameForm").reset()
 })
 
+.catch(err => console.log("error occured while adding game",err));
+}
+
 function deleteGame(id){
+
+    if(!confirm("are you sure you wish to proceed with this request?")) return;
+
         fetch(`${base_url}/${id}`,{
         method: "DELETE"})
         .then(res => res.json())
         .then(() => {console.log(`you have successfuly deleted ${id}`)
-        fetch(base_url)
-        .then(res => res.json())
-        .then(updatedData => myGames(updatedData));
+        fetchGames();
+       
 })
         .catch(err => console.error("an error occured while deleting the game",err));
 }
 
+function postGame(id){
+    alert(`downloading ${id}`)
+}
